third_party/blink/web_tests/platform/generic/fast/dom/shadow/hostcontext-first-child-expected.txt

Test that :host-context(:first-child) is re-evaluated when :first-child changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(first.firstChild, null).backgroundColor is red
PASS getComputedStyle(first.firstChild, null).backgroundColor is not red
PASS successfullyParsed is true

TEST COMPLETE
